Move GroupPhase related entities to separate module
