+++ ippon/group_phase/group_phase_views_test.py
